move GlobalMessageHistory to a separate file document GlobalMessageHistory class
